Move compact_size decoding into primitives witness
The compact_size module provides a single function decode_unchecked
that is only used by the witnesses in primitives. Since encoding
has already been moved to CompactSizeEncoder, we can move this
decoding function and remove internals::compact_size entirely.

Move decode_unchecked from internals::compact_size to witness.rs
in primitives and remove compact_size module from internals.
