+++ internals/src/lib.rs
@@ -38,7 +38,6 @@ pub mod _export {
 
 pub mod array;
 pub mod array_vec;
-pub mod compact_size;
 pub mod const_tools;
 pub mod error;
 pub mod macros;
